requirements.txt
apache-beam[gcp]==2.24.0
google-cloud-firestore==1.9.0